Move more to stanza.resources.common and add some documentation
